+++ packages/block-library/src/playlist-track/index.js
@@ -2,6 +2,7 @@ import { playlistTrack as icon } from '@wordpress/icons';
 import initBlock from '../utils/init-block';
 import metadata from './block.json';
 import edit from './edit';
+import transforms from './transforms';
 
 const { name } = metadata;
 export { metadata, name };
@@ -9,6 +10,7 @@ export { metadata, name };
 export const settings = {
 	icon,
 	edit,
+	transforms,
 };
 
 export const init = () => initBlock( { name, metadata, settings } );